+++ pennylane/devices/qubit/simulate.py
@@ -24,6 +24,7 @@ from numpy.random import default_rng
 
 import pennylane as qml
 from pennylane.logging import debug_logger
+from pennylane.math import Interface
 from pennylane.measurements import (
     CountsMP,
     ExpectationMP,
@@ -62,6 +63,15 @@ INTERFACE_TO_LIKE = {
     "tf": "tensorflow",
     "tensorflow-autograph": "tensorflow",
     "tf-autograph": "tensorflow",
+    # map interface enum
+    Interface.AUTOGRAD: "autograd",
+    Interface.NUMPY: "numpy",
+    Interface.TORCH: "torch",
+    Interface.JAX: "jax",
+    Interface.JAX_JIT: "jax",
+    Interface.TF: "tensorflow",
+    Interface.TF_AUTOGRAPH: "tensorflow",
+    Interface.AUTO: None,
 }
