+++ pennylane/math/__init__.py
@@ -88,12 +88,21 @@ from .utils import (
     allequal,
     cast,
     cast_like,
+)
+from .interface import (
     convert_like,
+    Interface,
+    get_canonical_interface,
+    SupportedInterfaceUserInput,
+    jpc_interfaces,
+    SUPPORTED_INTERFACE_NAMES,
+    INTERFACE_MAP,
     get_deep_interface,
     get_interface,
     in_backprop,
-    is_abstract,
     requires_grad,
+    is_abstract,
+    _get_interface_name,
 )
 
 sum = ar.numpy.sum
