+++ pennylane/math/multi_dispatch.py
@@ -24,7 +24,8 @@ from autoray import numpy as np
 from numpy import ndarray
 
 from . import single_dispatch  # pylint:disable=unused-import
-from .utils import cast, cast_like, get_interface, requires_grad
+from .interface import get_interface, requires_grad
+from .utils import cast, cast_like
 
 
 # pylint:disable=redefined-outer-name
